🤖 Refactor(#45): add directory in store
